+++ Source/VanillaFurnitureProduction.cs
@@ -1,0 +1,29 @@
+﻿namespace Replace_Stuff_Compatibility
+{
+	public class VanillaFurnitureProduction: AbstractPatch
+	{
+		protected override string GetRequiredModNames() => "vanillaexpanded.vfeproduction";
+
+		protected override void AddItems()
+		{
+			var stoneCutter = GetDatabaseThing("TableStonecutter");
+			var electricStoneCutter = GetDatabaseThing("VFE_TableStonecutterElectric");
+
+			var fabricationBench = GetDatabaseThing("FabricationBench");
+			var assemblyBench = GetDatabaseThing("VFE_ComponentFabricationBench");
+
+			var butchersTable = GetDatabaseThing("TableButcher");
+			var electricButchersTable = GetDatabaseThing("VFE_TableButcherElectric");
+
+			var drugLab = GetDatabaseThing("DrugLab");
+			var electricDrugLab = GetDatabaseThing("VFE_TableDrugLabElectric");
+			
+			MultiModPatch.Smelters.Add(GetDatabaseThing("VFE_FueledSmelter"));
+
+			AddInterchangeableList(stoneCutter, electricStoneCutter);
+			AddInterchangeableList(fabricationBench, assemblyBench);
+			AddInterchangeableList(butchersTable, electricButchersTable);
+			AddInterchangeableList(drugLab, electricDrugLab);
+		}
+	}
+}